simplify param type infer
