+++ rules/TypeDeclaration/TypeInferer/ParamTypeInferer.php
@@ -6,24 +6,34 @@ namespace Rector\TypeDeclaration\TypeInferer;
 
 use PhpParser\Node;
 use PhpParser\Node\Param;
+use PhpParser\Node\Stmt\ClassMethod;
 use PHPStan\Type\MixedType;
 use PHPStan\Type\Type;
+use Rector\BetterPhpDocParser\PhpDocInfo\PhpDocInfoFactory;
+use Rector\NodeNameResolver\NodeNameResolver;
 use Rector\NodeTypeResolver\NodeTypeResolver;
 use Rector\TypeDeclaration\TypeAnalyzer\GenericClassStringTypeNormalizer;
-use Rector\TypeDeclaration\TypeInferer\ParamTypeInferer\FunctionLikeDocParamTypeInferer;
 
 final class ParamTypeInferer
 {
     public function __construct(
         private readonly GenericClassStringTypeNormalizer $genericClassStringTypeNormalizer,
-        private readonly FunctionLikeDocParamTypeInferer $functionLikeDocParamTypeInferer,
-        private readonly NodeTypeResolver $nodeTypeResolver
+        private readonly NodeTypeResolver $nodeTypeResolver,
+        private readonly PhpDocInfoFactory $phpDocInfoFactory,
+        private readonly NodeNameResolver $nodeNameResolver
     ) {
     }
 
-    public function inferParam(Param $param): Type
+    public function resolveClassMethodParamDocType(ClassMethod $classMethod, Param $param): Type
     {
-        $paramType = $this->functionLikeDocParamTypeInferer->inferParam($param);
+        $phpDocInfo = $this->phpDocInfoFactory->createFromNodeOrEmpty($classMethod);
+
+        $paramTypesByName = $phpDocInfo->getParamTypesByName();
+        if ($paramTypesByName === []) {
+            return new MixedType();
+        }
+
+        $paramType = $this->matchParamNodeFromDoc($paramTypesByName, $param);
         if ($paramType instanceof MixedType) {
             return new MixedType();
         }
@@ -37,5 +47,14 @@ final class ParamTypeInferer
         }
 
         return $inferedType;
+    }
+
+    /**
+     * @param Type[] $paramWithTypes
+     */
+    private function matchParamNodeFromDoc(array $paramWithTypes, Param $param): Type
+    {
+        $paramNodeName = '$' . $this->nodeNameResolver->getName($param->var);
+        return $paramWithTypes[$paramNodeName] ?? new MixedType();
     }
 }
